requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
requests==2.32.3
tqdm==4.66.5
PyYAML==6.0.2
psutil==6.0.0
Django==5.2.4
PyMySQL==1.1.1
asgiref==3.9.1
sqlparse==0.5.3
tzdata>=2025.2
typing_extensions==4.12.2
nest-asyncio==1.6.0
lebai-sdk
flask>=3.0.0